+++ api/mraa/common.h
@@ -50,6 +50,17 @@
                                       if (res != MRAA_SUCCESS) \
                                       return res;} while(0)
 
+/**
+ * Simple deprecated macro
+ */
+#ifdef __GNUC__
+#define DEPRECATED __attribute__((deprecated))
+#elif defined(_MSC_VER)
+#define DEPRECATED __declspec(deprecated)
+#else
+#pragma message("WARNING: You need to implement DEPRECATED for this compiler")
+#define DEPRECATED
+#endif
 
 /** @file
  *
